[#175142335] Connects Search Bank screen with store and remote data
